+++ packages/expo-modules-core/ios/Fabric/ExpoFabricViewObjC.h
@@ -54,6 +54,8 @@
 
 - (void)viewDidUpdateProps NS_SWIFT_UI_ACTOR;
 
+- (void)updateLayoutMetrics:(CGRect)frame oldFrame:(CGRect)oldFrame NS_SWIFT_UI_ACTOR;
+
 - (void)setShadowNodeSize:(float)width height:(float)height NS_SWIFT_UI_ACTOR;
 
 - (void)setStyleSize:(nullable NSNumber *)width height:(nullable NSNumber *)height NS_SWIFT_UI_ACTOR;